Moves namespaces correctly
