+++ AppleAuth/AppleAuthQuickLoginArgs.cs.meta
