+++ AppleAuth/Enums/AuthorizationErrorCode.cs.meta
