+++ AppleAuth/Enums/CredentialState.cs.meta
