+++ AppleAuth/Extensions/AppleErrorExtensions.cs.meta
