+++ AppleAuth/IAppleAuthManager.cs
@@ -1,16 +1,19 @@
+using AppleAuth.Enums;
+using AppleAuth.Interfaces;
 using System;
-using AppleAuth.IOS.Enums;
-using AppleAuth.IOS.Interfaces;
 
-namespace AppleAuth.IOS
+namespace AppleAuth
 {
     public interface IAppleAuthManager
     {
         bool IsCurrentPlatformSupported { get; }
-        
+
         void QuickLogin(AppleAuthQuickLoginArgs quickLoginArgs, Action<ICredential> successCallback, Action<IAppleError> errorCallback);
+
         void LoginWithAppleId(AppleAuthLoginArgs loginArgs, Action<ICredential> successCallback, Action<IAppleError> errorCallback);
+
         void GetCredentialState(string userId, Action<CredentialState> successCallback, Action<IAppleError> errorCallback);
+
         void SetCredentialsRevokedCallback(Action<string> credentialsRevokedCallback);
 
         void Update();
