+++ AppleAuth/IAppleAuthManager.cs.meta
