+++ AppleAuth/Interfaces/IAppleError.cs
@@ -1,4 +1,4 @@
-namespace AppleAuth.IOS.Interfaces
+namespace AppleAuth.Interfaces
 {
     public interface IAppleError
     {
