+++ AppleAuth/Interfaces/IAppleError.cs.meta
