+++ AppleAuth/Interfaces/IAppleIDCredential.cs.meta
