+++ AppleAuth/Interfaces/ICredential.cs.meta
