+++ AppleAuth/Interfaces/ICredentialStateResponse.cs.meta
