+++ AppleAuth/Interfaces/ILoginWithAppleIdResponse.cs
@@ -1,4 +1,4 @@
-namespace AppleAuth.IOS.Interfaces
+namespace AppleAuth.Interfaces
 {
     public interface ILoginWithAppleIdResponse
     {
