+++ AppleAuth/Interfaces/ILoginWithAppleIdResponse.cs.meta
