+++ AppleAuth/Interfaces/IPasswordCredential.cs.meta
